Updated hair type to int
There's well over 255 hair styles now and packets seem to use short. Let's just go with int.

+++ src/ChannelServer/Scripting/ScriptManager.Functions.cs
@@ -840,7 +840,7 @@ namespace Melia.Channel.Scripting
 
 			lua_settop(L, 0);
 
-			character.Hair = (byte)hairId;
+			character.Hair = hairId;
 			Send.ZC_UPDATED_PCAPPEARANCE(character);
 
 			return 0;
